requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
selenium==4.15.2
webdriver-manager==4.0.1
pyautogui==0.9.54
pywinauto==0.6.8
openai==1.3.0
python-dotenv==1.0.0
websockets==12.0
langchain==0.0.335
chromadb==0.4.18
